requirements.txt
colorgram.py==1.2.0
pillow==10.2.0
